api/requirements.txt
autopep8==1.6.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycodestyle==2.8.0
PyJWT==2.3.0
toml==0.10.2
Werkzeug==2.0.3
